+++ django_redshift_backend/base.py
@@ -15,7 +15,9 @@ import django
 from django.conf import settings
 from django.core.exceptions import FieldDoesNotExist
 from django.db.backends.base.introspection import FieldInfo
+from django.db.backends.base.schema import _is_relevant_relation
 from django.db.backends.base.validation import BaseDatabaseValidation
+from django.db.backends.ddl_references import Statement
 from django.db.backends.postgresql.base import (
     DatabaseFeatures as BasePGDatabaseFeatures,
     DatabaseWrapper as BasePGDatabaseWrapper,
@@ -46,6 +48,7 @@ class DatabaseFeatures(BasePGDatabaseFeatures):
     has_native_uuid_field = False
     supports_aggregate_filter_clause = False
     can_rollback_ddl = False
+    supports_combined_alters = False          # since django-1.8
 
 
 class DatabaseOperations(BasePGDatabaseOperations):
@@ -122,6 +125,9 @@ def _related_non_m2m_objects(old_field, new_field):
 class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
 
     sql_create_table = "CREATE TABLE %(table)s (%(definition)s) %(options)s"
+    if django.VERSION < (3,):
+        # to remove "USING %(column)s::%(type)s"
+        sql_alter_column_type = "ALTER COLUMN %(column)s TYPE %(type)s"
 
     @property
     def multiply_varchar_length(self):
@@ -285,39 +291,52 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
         if self.connection.features.connection_persists_old_columns:
             self.connection.close()
 
+    # BASED FROM https://github.com/django/django/blob/3.2.12/django/db/backends/base/schema.py#L611-L864
     def _alter_field(self, model, old_field, new_field, old_type, new_type,
                      old_db_params, new_db_params, strict=False):
-        """Actually perform a "physical" (non-ManyToMany) field update."""
-
+        """Perform a "physical" (non-ManyToMany) field update."""
         # Drop any FK constraints, we'll remake them later
         fks_dropped = set()
-        if old_field.remote_field and old_field.db_constraint:
+        if (
+            self.connection.features.supports_foreign_keys and
+            old_field.remote_field and
+            old_field.db_constraint
+        ):
             fk_names = self._constraint_names(model, [old_field.column], foreign_key=True)
             if strict and len(fk_names) != 1:
-                raise ValueError(
-                    "Found wrong number (%s) of foreign key constraints for %s.%s" %
-                    (len(fk_names), model._meta.db_table, old_field.column))
+                raise ValueError("Found wrong number (%s) of foreign key constraints for %s.%s" % (
+                    len(fk_names),
+                    model._meta.db_table,
+                    old_field.column,
+                ))
             for fk_name in fk_names:
                 fks_dropped.add((old_field.column,))
-                self.execute(self._delete_constraint_sql(
-                    self.sql_delete_fk, model, fk_name))
+                self.execute(self._delete_fk_sql(model, fk_name))
         # Has unique been removed?
-        if (old_field.unique and
-                (not new_field.unique or
-                 (not old_field.primary_key and new_field.primary_key))):
+        if old_field.unique and (not new_field.unique or self._field_became_primary_key(old_field, new_field)):
             # Find the unique constraint for this field
+            meta_constraint_names = {constraint.name for constraint in model._meta.constraints}
             constraint_names = self._constraint_names(
-                model, [old_field.column], unique=True)
+                model, [old_field.column], unique=True, primary_key=False,
+                exclude=meta_constraint_names,
+            )
             if strict and len(constraint_names) != 1:
-                raise ValueError(
-                    "Found wrong number (%s) of unique constraints for %s.%s" %
-                    (len(constraint_names), model._meta.db_table, old_field.column))
+                raise ValueError("Found wrong number (%s) of unique constraints for %s.%s" % (
+                    len(constraint_names),
+                    model._meta.db_table,
+                    old_field.column,
+                ))
             for constraint_name in constraint_names:
-                self.execute(self._delete_constraint_sql(
-                    self.sql_delete_unique, model, constraint_name))
-        # Drop incoming FK constraints if we're a primary key and things are going
-        # to change.
-        if old_field.primary_key and new_field.primary_key and old_type != new_type:
+                self.execute(self._delete_unique_sql(model, constraint_name))
+        # Drop incoming FK constraints if the field is a primary key or unique,
+        # which might be a to_field target, and things are going to change.
+        drop_foreign_keys = (
+            self.connection.features.supports_foreign_keys and (
+                (old_field.primary_key and new_field.primary_key) or
+                (old_field.unique and new_field.unique)
+            ) and old_type != new_type
+        )
+        if drop_foreign_keys:
             # '_meta.related_field' also contains M2M reverse fields, these
             # will be filtered out
             for _old_rel, new_rel in _related_non_m2m_objects(old_field, new_field):
@@ -325,43 +344,319 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
                     new_rel.related_model, [new_rel.field.column], foreign_key=True
                 )
                 for fk_name in rel_fk_names:
-                    self.execute(self._delete_constraint_sql(
-                        self.sql_delete_fk, new_rel.related_model, fk_name))
+                    self.execute(self._delete_fk_sql(new_rel.related_model, fk_name))
         # Removed an index? (no strict check, as multiple indexes are possible)
-        if (old_field.db_index and
-                not new_field.db_index and
-                not old_field.unique and
-                not (not new_field.unique and old_field.unique)):
+        # Remove indexes if db_index switched to False or a unique constraint
+        # will now be used in lieu of an index. The following lines from the
+        # truth table show all True cases; the rest are False:
+        #
+        # old_field.db_index | old_field.unique | new_field.db_index | new_field.unique
+        # ------------------------------------------------------------------------------
+        # True               | False            | False              | False
+        # True               | False            | False              | True
+        # True               | False            | True               | True
+        if old_field.db_index and not old_field.unique and (not new_field.db_index or new_field.unique):
             # Find the index for this field
-            index_names = self._constraint_names(model, [old_field.column], index=True)
+            meta_index_names = {index.name for index in model._meta.indexes}
+            # Retrieve only BTREE indexes since this is what's created with
+            # db_index=True.
+            index_names = self._constraint_names(
+                model, [old_field.column], index=True, type_=Index.suffix,
+                exclude=meta_index_names,
+            )
             for index_name in index_names:
-                self.execute(self._delete_constraint_sql(
-                    self.sql_delete_index, model, index_name))
+                # The only way to check if an index was created with
+                # db_index=True or with Index(['field'], name='foo')
+                # is to look at its name (refs #28053).
+                self.execute(self._delete_index_sql(model, index_name))
         # Change check constraints?
         if old_db_params['check'] != new_db_params['check'] and old_db_params['check']:
+            meta_constraint_names = {constraint.name for constraint in model._meta.constraints}
             constraint_names = self._constraint_names(
-                model, [old_field.column], check=True)
+                model, [old_field.column], check=True,
+                exclude=meta_constraint_names,
+            )
             if strict and len(constraint_names) != 1:
-                raise ValueError(
-                    "Found wrong number (%s) of check constraints for %s.%s" %
-                    (len(constraint_names), model._meta.db_table, old_field.column))
+                raise ValueError("Found wrong number (%s) of check constraints for %s.%s" % (
+                    len(constraint_names),
+                    model._meta.db_table,
+                    old_field.column,
+                ))
             for constraint_name in constraint_names:
-                self.execute(self._delete_constraint_sql(
-                    self.sql_delete_check, model, constraint_name))
+                self.execute(self._delete_check_sql(model, constraint_name))
         # Have they renamed the column?
         if old_field.column != new_field.column:
-            self.execute(self._rename_field_sql(
-                model._meta.db_table, old_field, new_field, new_type))
+            self.execute(self._rename_field_sql(model._meta.db_table, old_field, new_field, new_type))
+            # Rename all references to the renamed column.
+            for sql in self.deferred_sql:
+                if isinstance(sql, Statement):
+                    sql.rename_column_references(model._meta.db_table, old_field.column, new_field.column)
         # Next, start accumulating actions to do
         actions = []
+        null_actions = []
         post_actions = []
-
+        # Collation change?
+        old_collation = getattr(old_field, 'db_collation', None)
+        new_collation = getattr(new_field, 'db_collation', None)
+        if old_collation != new_collation:
+            # Collation change handles also a type change.
+            fragment = self._alter_column_collation_sql(model, new_field, new_type, new_collation)
+            actions.append(fragment)
+        # Type change?
+        elif old_type != new_type:
+            fragment, other_actions = self._alter_column_type_sql(model, old_field, new_field, new_type)
+            if fragment:  # ## Redshift: In some case, fragment will be empty.
+                actions.append(fragment)
+            post_actions.extend(other_actions)
         # When changing a column NULL constraint to NOT NULL with a given
         # default value, we need to perform 4 steps:
         #  1. Add a default for new incoming writes
         #  2. Update existing NULL rows with new default
         #  3. Replace NULL constraint with NOT NULL
         #  4. Drop the default again.
+        # Default change?
+        # ## original do alter ALTER COLUMN SET/DROP DEFAULT.
+        # ## Redshift Can't: https://github.com/jazzband/django-redshift-backend/issues/96
+        # needs_database_default = False
+        # if old_field.null and not new_field.null:
+        #     old_default = self.effective_default(old_field)
+        #     new_default = self.effective_default(new_field)
+        #     if (
+        #         not self.skip_default_on_alter(new_field) and
+        #         old_default != new_default and
+        #         new_default is not None
+        #     ):
+        #         needs_database_default = True
+        #         actions.append(self._alter_column_default_sql(model, old_field, new_field))
+
+        # Nullability change?
+        if old_field.null != new_field.null:
+            # ## original BaseDatabaseSchemaEditor._alter_column_null_sql return only a fragment
+            # fragment = self._alter_column_null_sql(model, old_field, new_field)
+            # if fragment:
+            #     null_actions.append(fragment)
+            # ## Redshift use 4 steps null alternation
+            fragment, other_actions = self._alter_column_null_sqls(model, old_field, new_field)
+            actions.append(fragment)
+            null_actions.extend(other_actions)
+
+        # Only if we have a default and there is a change from NULL to NOT NULL
+        # ## original BaseDatabaseSchemaEditor._alter_table four_way_default_alteration
+        # four_way_default_alteration = (
+        #     new_field.has_default() and
+        #     (old_field.null and not new_field.null)
+        # )
+        # ## Redshift is always 4-way, this flag should be False because the null_actions handles it.
+        four_way_default_alteration = False
+        new_default = self.effective_default(new_field)  # not used, but for flake8
+
+        if actions or null_actions:
+            if not four_way_default_alteration:
+                # If we don't have to do a 4-way default alteration we can
+                # directly run a (NOT) NULL alteration
+                actions = actions + null_actions
+            # Combine actions together if we can (e.g. postgres)
+            if self.connection.features.supports_combined_alters and actions:
+                sql, params = tuple(zip(*actions))
+                actions = [(", ".join(sql), sum(params, []))]
+            # Apply those actions
+            for sql, params in actions:
+                self.execute(
+                    # ## original assumes only alters, so adding an ALTER TABLE clause
+                    # self.sql_alter_column % {
+                    #     "table": self.quote_name(model._meta.db_table),
+                    #     "changes": sql,
+                    # },
+                    # ## Redshift executes ADD and UPDATE on alter, so the sql is a complete sentence
+                    sql,
+                    params,
+                )
+            if four_way_default_alteration:
+                # Update existing rows with default value
+                self.execute(
+                    self.sql_update_with_default % {
+                        "table": self.quote_name(model._meta.db_table),
+                        "column": self.quote_name(new_field.column),
+                        "default": "%s",
+                    },
+                    [new_default],
+                )
+                # Since we didn't run a NOT NULL change before we need to do it
+                # now
+                for sql, params in null_actions:
+                    self.execute(
+                        self.sql_alter_column % {
+                            "table": self.quote_name(model._meta.db_table),
+                            "changes": sql,
+                        },
+                        params,
+                    )
+        if post_actions:
+            for sql, params in post_actions:
+                self.execute(sql, params)
+        # If primary_key changed to False, delete the primary key constraint.
+        if old_field.primary_key and not new_field.primary_key:
+            self._delete_primary_key(model, strict)
+        # Added a unique?
+        if self._unique_should_be_added(old_field, new_field):
+            self.execute(self._create_unique_sql(model, [new_field.column]))
+        # Added an index? Add an index if db_index switched to True or a unique
+        # constraint will no longer be used in lieu of an index. The following
+        # lines from the truth table show all True cases; the rest are False:
+        #
+        # old_field.db_index | old_field.unique | new_field.db_index | new_field.unique
+        # ------------------------------------------------------------------------------
+        # False              | False            | True               | False
+        # False              | True             | True               | False
+        # True               | True             | True               | False
+
+        # ## original BasePGDatabaseSchemaEditor._alter_field has CREATE INDEX
+        # ## Redshift doesn't support it.
+        # https://docs.aws.amazon.com/redshift/latest/dg/c_unsupported-postgresql-features.html
+        # if (not old_field.db_index or old_field.unique) and new_field.db_index and not new_field.unique:
+        #     self.execute(self._create_index_sql(model, fields=[new_field]))
+
+        # Type alteration on primary key? Then we need to alter the column
+        # referring to us.
+        rels_to_update = []
+        if drop_foreign_keys:
+            rels_to_update.extend(_related_non_m2m_objects(old_field, new_field))
+        # Changed to become primary key?
+        if self._field_became_primary_key(old_field, new_field):
+            # Make the new one
+            self.execute(self._create_primary_key_sql(model, new_field))
+            # Update all referencing columns
+            rels_to_update.extend(_related_non_m2m_objects(old_field, new_field))
+        # Handle our type alters on the other end of rels from the PK stuff above
+        for old_rel, new_rel in rels_to_update:
+            rel_db_params = new_rel.field.db_parameters(connection=self.connection)
+            rel_type = rel_db_params['type']
+            fragment, other_actions = self._alter_column_type_sql(
+                new_rel.related_model, old_rel.field, new_rel.field, rel_type
+            )
+            self.execute(
+                self.sql_alter_column % {
+                    "table": self.quote_name(new_rel.related_model._meta.db_table),
+                    "changes": fragment[0],
+                },
+                fragment[1],
+            )
+            for sql, params in other_actions:
+                self.execute(sql, params)
+        # Does it have a foreign key?
+        if (self.connection.features.supports_foreign_keys and new_field.remote_field and
+                (fks_dropped or not old_field.remote_field or not old_field.db_constraint) and
+                new_field.db_constraint):
+            self.execute(self._create_fk_sql(model, new_field, "_fk_%(to_table)s_%(to_column)s"))
+        # Rebuild FKs that pointed to us if we previously had to drop them
+        if drop_foreign_keys:
+            for rel in new_field.model._meta.related_objects:
+                if _is_relevant_relation(rel, new_field) and rel.field.db_constraint:
+                    self.execute(self._create_fk_sql(rel.related_model, rel.field, "_fk"))
+        # Does it have check constraints we need to add?
+        if old_db_params['check'] != new_db_params['check'] and new_db_params['check']:
+            constraint_name = self._create_index_name(model._meta.db_table, [new_field.column], suffix='_check')
+            self.execute(self._create_check_sql(model, constraint_name, new_db_params['check']))
+
+        # ## original BasePGDatabaseSchemaEditor._alter_field drop default here
+        # ## Redshift doesn't support DROP DEFAULT.
+        # Drop the default if we need to
+        # (Django usually does not use in-database defaults)
+        # if needs_database_default:
+        #     changes_sql, params = self._alter_column_default_sql(model, old_field, new_field, drop=True)
+        #     sql = self.sql_alter_column % {
+        #         "table": self.quote_name(model._meta.db_table),
+        #         "changes": changes_sql,
+        #     }
+        #     self.execute(sql, params)
+
+        # Reset connection if required
+        if self.connection.features.connection_persists_old_columns:
+            self.connection.close()
+
+    def _alter_column_with_recreate(self, model, old_field, new_field):
+        """
+        To change column type or default, We need this migration sequence:
+
+        1. Add new column with temporary name
+        2. Migrate values from original column to temprary column
+        3. Drop old column
+        4. Rename temporary column name to original column name
+        """
+        new_default = self.effective_default(new_field)
+
+        fragment = ('', [])
+        actions = []
+
+        # ## ALTER TABLE <table> ADD COLUMN 'tmp' <type> DEFAULT <value>
+        definition, params = self.column_sql(model, new_field, include_default=True)
+        new_defaults = [new_default] if new_default is not None else []
+        fragment = (
+            self.sql_create_column % {
+                "table": self.quote_name(model._meta.db_table),
+                "column": self.quote_name(new_field.column + "_tmp"),
+                "definition": definition,
+            },
+            new_defaults
+        )
+        # ## UPDATE <table> SET 'tmp' = <orig column>
+        actions.append((
+            "UPDATE %(table)s SET %(new_column)s = %(old_column)s" % {
+                "table": model._meta.db_table,
+                "new_column": self.quote_name(new_field.column + "_tmp"),
+                "old_column": self.quote_name(new_field.column),
+            }, []
+        ))
+        # ## ALTER TABLE <table> DROP COLUMN <orig column>
+        actions.append((
+            self.sql_delete_column % {
+                "table": model._meta.db_table,
+                "column": self.quote_name(new_field.column),
+            }, [],
+        ))
+        # ## ALTER TABLE <table> RENAME COLUMN 'tmp' <orig column>
+        actions.append((
+            self.sql_rename_column % {
+                "table": model._meta.db_table,
+                "old_column": self.quote_name(new_field.column + "_tmp"),
+                "new_column": self.quote_name(new_field.column),
+            }, []
+        ))
+
+        return fragment, actions
+
+    # BASED FROM https://github.com/django/django/blob/3.2.12/django/db/backends/base/schema.py#L866-L886
+    # postgres/schema.py doesn't have `_alter_column_null_sql` method.
+    def _alter_column_null_sqls(self, model, old_field, new_field):
+        """
+        Hook to specialize column null alteration.
+
+        Return a [(sql, params), ...] fragment to set a column to null or non-null
+        as required by new_field, or None if no changes are required.
+        """
+        # ## original BaseDatabaseSchemaEditor._alter_column_null_sql return a sql fragment
+        # ## Redshift needs 4 step alter
+        return self._alter_column_with_recreate(model, old_field, new_field)
+
+    # override to avoid `USING %(column)s::%(type)s` on postgres/schema.py
+    def _alter_column_type_sql(self, model, old_field, new_field, new_type):
+        # """
+        # Hook to specialize column type alteration for different backends,
+        # for cases when a creation type is different to an alteration type
+        # (e.g. SERIAL in PostgreSQL, PostGIS fields).
+
+        # Return a two-tuple of: an SQL fragment of (sql, params) to insert into
+        # an ALTER TABLE statement and a list of extra (sql, params) tuples to
+        # run once the field is altered.
+        # """
+        old_db_params = old_field.db_parameters(connection=self.connection)
+        old_type = old_db_params['type']
+        old_default = self.effective_default(old_field)
+        new_default = self.effective_default(new_field)
+
+        fragment = ('', [])
+        actions = []
+
         # Default change?
         old_default = self.effective_default(old_field)
         new_default = self.effective_default(new_field)
@@ -371,6 +666,10 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
             not self.skip_default(new_field)
         )
         # column type changed (NG pattern)
+        # MEMO: for a future, another procedure to ALTER:
+        #   1. once remove UNIQUE, PKEY, FK
+        #   2. alter column
+        #   3. add UNIQUE, PKEY, FK again
         if (old_type != new_type and (
                 old_field.unique or new_field.unique or
                 old_field.primary_key or new_field.primary_key or
@@ -378,12 +677,13 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
             # ## WARNING: Redshift Can't ADD/ALTER/DROP COLUMN with UNIQUE, PRIMARY KEY or FOREIGN KEY.
             # https://github.com/jazzband/django-redshift-backend/issues/96
             # https://docs.aws.amazon.com/en_us/redshift/latest/dg/r_ALTER_TABLE.html
-            fragment, other_actions = self._alter_column_type_sql(
-                model, old_field, new_field, new_type
-            )
+            _fragment = self.sql_alter_column_type % {
+                "column": self.quote_name(new_field.column),
+                "type": new_type,
+            }
             sql = self.sql_alter_column % {
                 "table": self.quote_name(model._meta.db_table),
-                "changes": fragment[0],
+                "changes": _fragment,
             }
             msg = '\n'.join([
                 "!!! WARNING: SKIPPED ALTER COLUMN SIZE WITH PRIMARY KEY, UNIQUE, FOREIGN KEY !!!",
@@ -398,6 +698,10 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
             msgfmt = dict(sql=sql, o=old_field, n=new_field, otype=old_type, ntype=new_type)
             logger.warning(msg.format(**msgfmt))
             sys.stdout.write('\n\n' + msg.format(**msgfmt) + '\n\n')
+            # actions.append(['', []])
+            # self.sql_alter_column = ''  # cancel  # FIXME comment and impl
+            # FIXME: need UnitTest
+            breakpoint()
 
         # Size is changed
         elif (type(old_field) == type(new_field) and
@@ -406,153 +710,22 @@ class DatabaseSchemaEditor(BasePGDatabaseSchemaEditor):
               old_field.max_length != new_field.max_length):
             # if shrink size as `old_field.max_length > new_field.max_length` and
             # larger data in database, this change will raise exception.
-            fragment, other_actions = self._alter_column_type_sql(
-                model, old_field, new_field, new_type
-            )
-            actions.append((
+            fragment = (
                 self.sql_alter_column % {
                     "table": self.quote_name(model._meta.db_table),
-                    "changes": fragment[0],
+                    "changes": self.sql_alter_column_type % {
+                        "column": self.quote_name(new_field.column),
+                        "type": new_type,
+                    }
                 },
-                fragment[1]
-            ))
+                []
+            )
 
         # Type or default is changed?
         elif (old_type != new_type) or needs_database_default:
-            # ## To change column type or default, We need this migration sequence:
-            # ##
-            # ## 1. Add new column with temporary name
-            # ## 2. Migrate values from original column to temprary column
-            # ## 3. Drop old column
-            # ## 4. Rename temporary column name to original column name
+            fragment, actions = self._alter_column_with_recreate(model, old_field, new_field)
 
-            # ## ALTER TABLE <table> ADD COLUMN 'tmp' <type> DEFAULT <value>
-            definition, params = self.column_sql(model, new_field, include_default=True)
-            new_defaults = [new_default] if new_default is not None else []
-            actions.append((
-                self.sql_create_column % {
-                    "table": self.quote_name(model._meta.db_table),
-                    "column": self.quote_name(new_field.column + "_tmp"),
-                    "definition": definition,
-                },
-                new_defaults
-            ))
-            # ## UPDATE <table> SET 'tmp' = <orig column>
-            actions.append((
-                "UPDATE %(table)s SET %(new_column)s = %(old_column)s" % {
-                    "table": model._meta.db_table,
-                    "new_column": self.quote_name(new_field.column + "_tmp"),
-                    "old_column": self.quote_name(new_field.column),
-                }, []
-            ))
-            # ## ALTER TABLE <table> DROP COLUMN <orig column>
-            actions.append((
-                self.sql_delete_column % {
-                    "table": model._meta.db_table,
-                    "column": self.quote_name(new_field.column),
-                }, [],
-            ))
-            # ## ALTER TABLE <table> RENAME COLUMN 'tmp' <orig column>
-            actions.append((
-                self.sql_rename_column % {
-                    "table": model._meta.db_table,
-                    "old_column": self.quote_name(new_field.column + "_tmp"),
-                    "new_column": self.quote_name(new_field.column),
-                }, []
-            ))
-
-        # Apply those actions
-        for sql, params in actions:
-            self.execute(sql, params)
-        if post_actions:
-            for sql, params in post_actions:
-                self.execute(sql, params)
-        # Added a unique?
-        if ((not old_field.unique and new_field.unique) or
-           (old_field.primary_key and not new_field.primary_key and new_field.unique)):
-            self.execute(self._create_unique_sql(model, [new_field.column]))
-
-        # ## original BasePGDatabaseSchemaEditor._alter_field has CREATE INDEX
-        # ## Redshift doesn't support it.
-
-        # Type alteration on primary key? Then we need to alter the column
-        # referring to us.
-        rels_to_update = []
-        if old_field.primary_key and new_field.primary_key and old_type != new_type:
-            rels_to_update.extend(_related_non_m2m_objects(old_field, new_field))
-        # Changed to become primary key?
-        # Note that we don't detect unsetting of a PK, as we assume another field
-        # will always come along and replace it.
-        if not old_field.primary_key and new_field.primary_key:
-            # First, drop the old PK
-            constraint_names = self._constraint_names(model, primary_key=True)
-            if strict and len(constraint_names) != 1:
-                raise ValueError("Found wrong number (%s) of PK constraints for %s" % (
-                    len(constraint_names),
-                    model._meta.db_table,
-                ))
-            for constraint_name in constraint_names:
-                self.execute(self._delete_constraint_sql(
-                    self.sql_delete_pk, model, constraint_name))
-            # Make the new one
-            self.execute(
-                self.sql_create_pk % {
-                    "table": self.quote_name(model._meta.db_table),
-                    "name": self.quote_name(self._create_index_name(
-                        model, [new_field.column], suffix="_pk")),
-                    "columns": self.quote_name(new_field.column),
-                }
-            )
-            # Update all referencing columns
-            rels_to_update.extend(_related_non_m2m_objects(old_field, new_field))
-        # Handle our type alters on the other end of rels from the PK stuff above
-        for old_rel, new_rel in rels_to_update:
-            rel_db_params = new_rel.field.db_parameters(connection=self.connection)
-            rel_type = rel_db_params['type']
-            fragment, other_actions = self._alter_column_type_sql(
-                new_rel.related_model._meta.db_table, old_rel.field, new_rel.field,
-                rel_type
-            )
-            self.execute(
-                self.sql_alter_column % {
-                    "table": self.quote_name(new_rel.related_model._meta.db_table),
-                    "changes": fragment[0],
-                },
-                fragment[1])
-            for sql, params in other_actions:
-                self.execute(sql, params)
-        # Does it have a foreign key?
-        if (new_field.remote_field and
-                (fks_dropped or
-                    not old_field.remote_field or
-                    not old_field.db_constraint) and
-                new_field.db_constraint):
-            self.execute(self._create_fk_sql(
-                model, new_field, "_fk_%(to_table)s_%(to_column)s"))
-        # Rebuild FKs that pointed to us if we previously had to drop them
-        if old_field.primary_key and new_field.primary_key and old_type != new_type:
-            for rel in new_field.model._meta.related_objects:
-                if not rel.many_to_many:
-                    self.execute(self._create_fk_sql(
-                        rel.related_model, rel.field, "_fk"))
-        # Does it have check constraints we need to add?
-        if old_db_params['check'] != new_db_params['check'] and new_db_params['check']:
-            self.execute(
-                self.sql_create_check % {
-                    "table": self.quote_name(model._meta.db_table),
-                    "name": self.quote_name(self._create_index_name(
-                        model, [new_field.column], suffix="_check")),
-                    "column": self.quote_name(new_field.column),
-                    "check": new_db_params['check'],
-                }
-            )
-
-        # ## original BasePGDatabaseSchemaEditor._alter_field drop default here
-        # ## Redshift doesn't support DROP DEFAULT.
-
-        # Reset connection if required
-        if self.connection.features.connection_persists_old_columns:
-            self.connection.close()
+        return fragment, actions
 
     def _get_create_options(self, model):
         """